Change: Completed refactoring of monolith to crates

+++ jcbench.app/crates/benchmark_submission/src/lib.rs
@@ -10,14 +10,14 @@ pub struct BenchmarkRequest {
     pub os_name: String,
     pub benchmark_name: String,
     pub benchmark_threading_model: String,
-    pub benchmark_api_version: u32,
+    pub benchmark_api_version: usize,
     pub cpu_name: String,
     pub cpu_architecture: String,
     pub cpu_cores: usize,
     pub score: u32
 }
 
-pub fn get_benchmark_request(selected_benchmark_name: String, selected_benchmark_api: u32, benchmark_settings: BenchmarkSettings, benchmark_result: u32) -> BenchmarkRequest {
+pub fn get_benchmark_request(selected_benchmark_name: String, selected_benchmark_api: usize, benchmark_settings: BenchmarkSettings, benchmark_result: u32) -> BenchmarkRequest {
     let cpu_info = CPUInfo::get_cpu_info();
 
     return BenchmarkRequest {
